requirements.txt
rich
questionary
numpy
llama-cpp-python
